Add matching service wrapper
Introduce a background service layer around the matching engine so orders can be queued and processed asynchronously. This keeps the core engine reusable now and leaves a clean seam for Kafka integration later.

+++ include/core/Types.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
  
-typedef uint8_t OrderId;
+typedef uint64_t OrderId;
 typedef double Price;
 
 enum class Side : uint8_t {
